+++ CHANGELOG.md
@@ -7,6 +7,11 @@ All notable changes to this project will be documented in this file.
 
 ## Unreleased
 
+### Added
+
+-   Multi spawn locations
+    -   When moving a shape to a new location that has multiple spawn locations, a box will appear to choose the desired spawn zone
+
 # [0.22.2] - 2020-08-28
 
 ### Fixed